Update Women in Blockchain Conference
Women in Blockchain Conference is an initiative by the Blockchain and Web3 Community Network (BWCN), dedicated to amplifying the voices and participation of women in the rapidly evolving blockchain and Web3 space. This empowering event brings together thought leaders, developers, entrepreneurs, students, and advocates to explore the intersection of gender, technology, and financial inclusion.