Update BWCN-XSpaces & Webnars
X Spaces and Webinars are virtual discussion platforms created to explore and unpack the most pertinent issues shaping the blockchain and Web3 space in Africa. From deep dives into the technology itself to critical conversations around policy, regulation, and adoption, these sessions bring together developers, policymakers, founders, and community leaders. The goal is to foster open, inclusive dialogue that informs, educates, and drives meaningful progress—ensuring that Africans are not just consumers of emerging tech, but active participants and shapers of it.