Update BWCN-Capstone Projects
The Capstone Project is a culmination of knowledge and skills gained throughout a program, providing participants with an opportunity to apply their learning to a real-world problem. This project challenges individuals or teams to design, develop, and present innovative solutions using blockchain, Web3, or AI technologies. Through collaboration, research, and hands-on development, the Capstone Project offers a platform for showcasing the ability to tackle complex challenges and make an impact in the evolving tech landscape. It serves as a final step in the journey of transforming theoretical knowledge into practical, actionable outcomes.